finished

+++ src/main/java/com/psaawkel/fibonacci/IFibonacci.java
@@ -2,6 +2,6 @@ package com.psaawkel.fibonacci;
 
 public interface IFibonacci {
 
-    int calculate(int factor);
+    long calculate(int factor);
 
 }
